release without tests

+++ lab_4/booking/src/main/java/org/example/service/impl/RoomServiceImpl.java
@@ -8,6 +8,7 @@ import org.example.dto.response.RoomWithHotelAmenitiesDTO;
 import org.example.entity.amenity.Amenity;
 import org.example.entity.apartment.Hotel;
 import org.example.entity.apartment.Room;
+import org.example.exception.ApartmentIsBookedException;
 import org.example.exception.RoomNotFoundException;
 import org.example.service.HotelService;
 import org.example.service.RoomService;
@@ -61,7 +62,7 @@ public class RoomServiceImpl implements RoomService {
 	@Transactional
 	public Room addRoomToHotel(RoomRequestDTO roomRequestDTO) {
 		Hotel hotel = hotelService.findHotelById(roomRequestDTO.getHotelId());
-		List<Amenity> amenities = amenityUtil.findAllById(roomRequestDTO.getAmenityIds());
+		List<Amenity> amenities = amenityUtil.getAmenities(roomRequestDTO.getAmenityIds());
 
 		Room room = new Room(roomRequestDTO.getName(), roomRequestDTO.getPrice(),
 				roomRequestDTO.getType(), amenities, roomRequestDTO.getMaxSpace(), hotel);
@@ -76,10 +77,29 @@ public class RoomServiceImpl implements RoomService {
 	@Transactional
 	public void deleteRoomFromHotel(Long roomId) {
 		Room room = findRoomById(roomId);
+
+		if (amenityUtil.existsByApartment(room)) {
+			throw new ApartmentIsBookedException(HttpStatus.BAD_REQUEST.value(), "Cannot delete room because it is linked to a booking.");
+		}
+
 		Hotel hotel = room.getHotel();
 		hotel.getRooms().remove(room);
 		hotelService.save(hotel);
 		roomRepository.delete(room);
+	}
+
+	@Override
+	@Transactional
+	public void deleteAllRooms(List<Room> rooms) {
+		boolean hasBookedRooms = rooms.stream()
+				.anyMatch(amenityUtil::existsByApartment);
+
+		if (hasBookedRooms) {
+			throw new ApartmentIsBookedException(HttpStatus.BAD_REQUEST.value(),
+					"Cannot delete hotel because one or more rooms are linked to a booking.");
+		}
+
+		roomRepository.deleteAll(rooms);
 	}
 
 	@Override
@@ -88,10 +108,5 @@ public class RoomServiceImpl implements RoomService {
 		List<Amenity> amenities = amenityUtil.getAmenities(roomRequestDTO.getAmenityIds());
 		room.updateFields(roomRequestDTO, updaterField, amenities);
 		return roomRepository.save(room);
-	}
-
-	@Override
-	public void deleteAllRooms(List<Room> rooms) {
-		roomRepository.deleteAll(rooms);
 	}
 }
